+++ src/Expression/Value/Builder/StructuredValueBuilder.php
@@ -2,11 +2,11 @@
 
 declare(strict_types=1);
 
-namespace Yiisoft\Db\Expression\Builder;
+namespace Yiisoft\Db\Expression\Value\Builder;
 
-use Yiisoft\Db\Expression\Param;
+use Yiisoft\Db\Expression\Value\Param;
 use Yiisoft\Db\Constant\DataType;
-use Yiisoft\Db\Expression\StructuredExpression;
+use Yiisoft\Db\Expression\Value\StructuredValue;
 use Yiisoft\Db\Query\QueryInterface;
 use Yiisoft\Db\Schema\Data\JsonLazyArray;
 use Yiisoft\Db\Schema\Data\LazyArray;
@@ -19,25 +19,25 @@ use function json_encode;
 use const JSON_THROW_ON_ERROR;
 
 /**
- * Default expression builder for {@see StructuredExpression}. Builds an expression as a JSON.
+ * Default expression builder for {@see StructuredValue}. Builds an expression as a JSON.
  */
-final class StructuredExpressionBuilder extends AbstractStructuredExpressionBuilder
+final class StructuredValueBuilder extends AbstractStructuredValueBuilder
 {
-    protected function buildStringValue(string $value, StructuredExpression $expression, array &$params): string
+    protected function buildStringValue(string $value, StructuredValue $expression, array &$params): string
     {
         $param = new Param($value, DataType::STRING);
 
         return $this->queryBuilder->bindParam($param, $params);
     }
 
-    protected function buildSubquery(QueryInterface $query, StructuredExpression $expression, array &$params): string
+    protected function buildSubquery(QueryInterface $query, StructuredValue $expression, array &$params): string
     {
         [$sql, $params] = $this->queryBuilder->build($query, $params);
 
         return "($sql)";
     }
 
-    protected function buildValue(array|object $value, StructuredExpression $expression, array &$params): string
+    protected function buildValue(array|object $value, StructuredValue $expression, array &$params): string
     {
         $value = $this->prepareValues($value, $expression);
         $param = new Param(json_encode(array_values($value), JSON_THROW_ON_ERROR), DataType::STRING);