+++ src/Schema/Column/AbstractStructuredColumn.php
@@ -6,7 +6,7 @@ namespace Yiisoft\Db\Schema\Column;
 
 use Yiisoft\Db\Constant\ColumnType;
 use Yiisoft\Db\Expression\ExpressionInterface;
-use Yiisoft\Db\Expression\StructuredExpression;
+use Yiisoft\Db\Expression\Value\StructuredValue;
 
 /**
  * Represents an abstract structured column.
@@ -57,6 +57,6 @@ abstract class AbstractStructuredColumn extends AbstractColumn
             return $value;
         }
 
-        return new StructuredExpression($value, $this);
+        return new StructuredValue($value, $this);
     }
 }