+++ tests/Db/Expression/Function/MultiOperandFunctionTest.php
@@ -7,14 +7,14 @@ namespace Yiisoft\Db\Tests\Db\Expression\Function;
 use PHPUnit\Framework\Attributes\DataProvider;
 use PHPUnit\Framework\TestCase;
 use Yiisoft\Db\Constant\DataType;
-use Yiisoft\Db\Expression\ArrayExpression;
+use Yiisoft\Db\Expression\Value\ArrayValue;
 use Yiisoft\Db\Expression\Function\ArrayMerge;
 use Yiisoft\Db\Expression\Function\Greatest;
 use Yiisoft\Db\Expression\Function\Least;
 use Yiisoft\Db\Expression\Function\Longest;
 use Yiisoft\Db\Expression\Function\MultiOperandFunction;
 use Yiisoft\Db\Expression\Function\Shortest;
-use Yiisoft\Db\Expression\Param;
+use Yiisoft\Db\Expression\Value\Param;
 use Yiisoft\Db\Tests\Support\TestTrait;
 
 final class MultiOperandFunctionTest extends TestCase
@@ -29,9 +29,9 @@ final class MultiOperandFunctionTest extends TestCase
         return [
             ArrayMerge::class => [ArrayMerge::class, [
                 [[1, 2, 3]],
-                [new ArrayExpression([1, 2, 3])],
+                [new ArrayValue([1, 2, 3])],
                 [$query],
-                [[1, 2, 3], '[1,2,3]', new ArrayExpression([1, 2, 3]), $query],
+                [[1, 2, 3], '[1,2,3]', new ArrayValue([1, 2, 3]), $query],
             ]],
             Greatest::class => [Greatest::class, [
                 [1],