+++ tests/Db/Expression/Value/JsonValueTest.php
@@ -2,12 +2,12 @@
 
 declare(strict_types=1);
 
-namespace Yiisoft\Db\Tests\Db\Expression;
+namespace Yiisoft\Db\Tests\Db\Expression\Value;
 
 use ArrayIterator;
 use PHPUnit\Framework\Attributes\DataProvider;
 use PHPUnit\Framework\TestCase;
-use Yiisoft\Db\Expression\JsonExpression;
+use Yiisoft\Db\Expression\Value\JsonValue;
 use Yiisoft\Db\Query\Query;
 use Yiisoft\Db\Schema\Data\JsonLazyArray;
 use Yiisoft\Db\Tests\Support\TestTrait;
@@ -15,7 +15,7 @@ use Yiisoft\Db\Tests\Support\TestTrait;
 /**
  * @group db
  */
-final class JsonExpressionTest extends TestCase
+final class JsonValueTest extends TestCase
 {
     use TestTrait;
 
@@ -39,9 +39,9 @@ final class JsonExpressionTest extends TestCase
         mixed $value,
         string|null $type = null
     ): void {
-        $expression = new JsonExpression($value, $type);
+        $expression = new JsonValue($value, $type);
 
-        $this->assertSame($value, $expression->getValue());
-        $this->assertSame($type, $expression->getType());
+        $this->assertSame($value, $expression->value);
+        $this->assertSame($type, $expression->type);
     }
 }