+++ tests/Provider/QueryBuilderProvider.php
@@ -8,23 +8,24 @@ use ArrayIterator;
 use DateTimeImmutable;
 use DateTimeZone;
 use Yiisoft\Db\Constant\DataType;
-use Yiisoft\Db\Expression\ArrayExpression;
+use Yiisoft\Db\Expression\Statement\When;
+use Yiisoft\Db\Expression\Value\ArrayValue;
 use Yiisoft\Db\Expression\Function\ArrayMerge;
-use Yiisoft\Db\Expression\Param;
+use Yiisoft\Db\Expression\Value\Param;
 use Yiisoft\Db\Constant\ColumnType;
 use Yiisoft\Db\Constant\IndexType;
 use Yiisoft\Db\Constant\PseudoType;
 use Yiisoft\Db\Constant\ReferentialAction;
 use Yiisoft\Db\Constraint\ForeignKey;
-use Yiisoft\Db\Expression\CaseExpression;
-use Yiisoft\Db\Expression\ColumnName;
+use Yiisoft\Db\Expression\Statement\CaseX;
+use Yiisoft\Db\Expression\Value\ColumnName;
 use Yiisoft\Db\Expression\Expression;
 use Yiisoft\Db\Expression\Function\Greatest;
 use Yiisoft\Db\Expression\Function\Least;
 use Yiisoft\Db\Expression\Function\Longest;
 use Yiisoft\Db\Expression\Function\Shortest;
-use Yiisoft\Db\Expression\JsonExpression;
-use Yiisoft\Db\Expression\Value;
+use Yiisoft\Db\Expression\Value\JsonValue;
+use Yiisoft\Db\Expression\Value\Value;
 use Yiisoft\Db\Expression\Value\DateTimeValue;
 use Yiisoft\Db\Query\Query;
 use Yiisoft\Db\QueryBuilder\Condition\All;
@@ -1161,7 +1162,7 @@ class QueryBuilderProvider
             'json expression' => [
                 'json_type',
                 [
-                    'json_col' => new JsonExpression(['c' => 1, 'd' => 2]),
+                    'json_col' => new JsonValue(['c' => 1, 'd' => 2]),
                 ],
                 [],
                 static::replaceQuotes(
@@ -1725,8 +1726,8 @@ class QueryBuilderProvider
             [new ArrayIterator([0, 1, 2, 7]), 1],
             'null' => [[null], 1],
             'expression' => [new Expression("'[0,1,2,7]'"), 1],
-            'json expression' => [new JsonExpression([0, 1, 2, 7]), 1],
-            'query expression' => [(new Query(static::getDb()))->select(new JsonExpression([0, 1, 2, 7])), 1],
+            'json expression' => [new JsonValue([0, 1, 2, 7]), 1],
+            'query expression' => [(new Query(static::getDb()))->select(new JsonValue([0, 1, 2, 7])), 1],
         ];
     }
 
@@ -2009,32 +2010,38 @@ class QueryBuilderProvider
         ];
     }
 
-    public static function caseExpressionBuilder(): array
+    public static function caseXBuilder(): array
     {
         return [
             'with case expression' => [
-                (new CaseExpression('(1 + 2)'))
-                    ->addWhen(1, 1)
-                    ->addWhen(2, new Expression('2'))
-                    ->addWhen(3, '(2 + 1)')
-                    ->else($param = new Param(4, DataType::INTEGER)),
+                new CaseX(
+                    '(1 + 2)',
+                    when1: new When(1, 1),
+                    when2: new When(2, new Expression('2')),
+                    when3: new When(3, '(2 + 1)'),
+                    else: $param = new Param(4, DataType::INTEGER),
+                ),
                 'CASE (1 + 2) WHEN 1 THEN 1 WHEN 2 THEN 2 WHEN 3 THEN (2 + 1) ELSE :qp0 END',
                 [':qp0' => $param],
                 3,
             ],
             'without case expression' => [
-                (new CaseExpression())
-                    ->addWhen(['=', 'column_name', 1], $paramA = new Param('a', DataType::STRING))
-                    ->addWhen(
+                new CaseX(
+                    when1: new When(['=', 'column_name', 1], new Value('a')),
+                    when2: new When(
                         static::replaceQuotes('[[column_name]] = 2'),
-                        (new Query(self::getDb()))->select($paramB = new Param('b', DataType::STRING))
+                        (new Query(self::getDb()))->select($param = new Param('b', DataType::STRING))
                     ),
+                ),
                 static::replaceQuotes(
                     <<<SQL
                     CASE WHEN [[column_name]] = 1 THEN :qp0 WHEN [[column_name]] = 2 THEN (SELECT :qp1) END
                     SQL
                 ),
-                [':qp0' => $paramA, ':qp1' => $paramB],
+                [
+                    ':qp0' => new Param('a', DataType::STRING),
+                    ':qp1' => $param,
+                ],
                 'b',
             ],
         ];
@@ -2180,7 +2187,7 @@ class QueryBuilderProvider
         return [[
             [
                 'id' => 1,
-                'array_col' => new ArrayExpression([1, 2, 3]),
+                'array_col' => new ArrayValue([1, 2, 3]),
                 'greatest_col' => 10,
                 'least_col' => 10,
                 'longest_col' => 'longest',
@@ -2188,7 +2195,7 @@ class QueryBuilderProvider
             ],
             [
                 'id' => 1,
-                'array_col' => new ArrayExpression([3, 4, 5]),
+                'array_col' => new ArrayValue([3, 4, 5]),
                 'greatest_col' => 5,
                 'least_col' => 5,
                 'longest_col' => 'short',